tests/submission/MOSQUITO/event-source/readystate-each-state-connection.html

<!DOCTYPE html>
<html>
<head>
	<title>The readyState attribute represents the state of the connection. It can have the following values: 0 (CONNECTING), 1 (OPEN), 2 (CLOSED):</title>
	<meta charset="utf-8" />
  	<meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
	  <link rel="author" title="Fabrice Clari" href="mailto:f.clari@inno-group.com">
	  <link rel="author" title="Dimitri Bocquet" href="mailto:Dimitri.Bocquet@mosquito-fp7.eu">
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="/resources/WebIDLParser.js"></script>
	<script src="/resources/idlharness.js"></script>
</head>
<body>
	<h1>The readyState attribute represents the state of the connection. It can have the following values: 0 (CONNECTING), 1 (OPEN), 2 (CLOSED):</h1>
	<div id="log"></div>
  <script>
  setup({timeout:6000});
    var source = new EventSource('http://html5.mosquito-fp7.eu/messages/t0016');
    var tconnect = async_test("The readyState attribute is 0 (CONNECTING)");
    var topen = async_test("The readyState attribute is 1 (OPEN)");
    var tclose = async_test("The readyState attribute is 2 (CLOSED)");
    
    tconnect.step(function() {assert_equals(source.readyState, EventSource.CONNECTING)});
    tconnect.done();
    
    source.addEventListener('message', function(e){
	}, false);
    source.addEventListener('open', function(e){ 
    	/* Connection opened. */
    	topen.step(function() {assert_equals(source.readyState, EventSource.OPEN)});
		topen.done();
		source.close();
		tclose.step(function() {assert_equals(source.readyState, EventSource.CLOSED)});
		tclose.done();
    	}, false);
    source.addEventListener('error', function(e){
    	/* Connection closed. */
    	}, false);
    
  </script>
  </div>
</body>
</html>
